configure JWT
